extract several controller from one
